lid/requirements/requirements.txt
einops==0.4.1
huggingface-hub==0.4.0
hydra-core==1.0.7
kenlm @ https://github.com/kpu/kenlm/archive/master.zip
librosa==0.8.1
scikit-learn==1.0.2
SoundFile==0.10.3.post1
sox==1.4.1
tensorboard==2.9.1
tensorboardX==2.4
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.6.2
torchvision==0.12.0
tqdm==4.62.3
transformers==4.15.0
wandb==0.12.20